+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Change log
 ### 6.13.1 (7/4/2019 - )
+* db: add Query.groupBy
 
 ### 6.13.0 (6/26/2019 - 7/1/2019)
 * search: update es to 7.2.0